refactor: add `NEW_VIDEO_SCHEMA`
